Fixed reader's logic, now each reader reads a new line for himself. No more loss of data. Feature: now works with unknown number of lines.

+++ fileReader/src/main/java/org/example/ExternalSorter.java
@@ -7,35 +7,59 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+/*
+    TODO: Remove Readers which have already finished executing.
+          Remove the corresponding firstLines when a reader has finished its work.  firstLines saves every file's first line.
+          Break the while loop when bufferedReaders array reaches 0
+          Decrease slices when a reader finishes.
+          Introduce threads to splitting and sorting the main file to decrease time.
+ */
+
 public class ExternalSorter {
-    static int lines = 100000000;
+    static int lines = 0;
     static int maxElements = 100000;
+    static String sortFileDir = "C:\\csv\\100m.csv";
     static String fileNames = "sortedGeneration";
 
     public static void main(String[] args) throws IOException {
         long begin = System.currentTimeMillis();
-        int slices = (int) Math.ceil((double) lines / maxElements);
-        try (FileReader fileToSort = new FileReader("C:\\csv\\100m.csv");
+        int slices;
+
+        FileReader lineFile = new FileReader(sortFileDir);
+        BufferedReader lineCounter = new BufferedReader(lineFile);
+        String line;
+        //Skips first line due to invalid characters
+        while (lineCounter.readLine() != null) {
+            lines++;
+        }
+        System.out.println(lines);
+        lineFile.close();
+        lineCounter.close();
+        try (FileReader fileToSort = new FileReader(sortFileDir);
              BufferedReader buffer = new BufferedReader(fileToSort)) {
-            String line;
-            line = buffer.readLine();
+            slices = (int) Math.ceil((double) lines / maxElements);
             String[] elements = new String[maxElements];
             int sliceNumber = 0;
+            int lastFile = maxElements;
             for (int i = 0; i < slices; i++) {
-                for (int j = 0; j < maxElements; j++) {
+                for (int j = 0; j < lastFile; j++) {
                     line = buffer.readLine();
                     elements[j] = line;
                 }
-                mergesort(elements);
+                    mergesort(elements);
                 //Write slices
                 writeSortedFile(elements, sliceNumber);
                 sliceNumber++;
-                elements = new String[maxElements];
+                if (i >= (slices - 2) && (lines%maxElements != 0)) {
+                    elements = new String[lines % maxElements];
+                    lastFile = lines%maxElements;
+                } else {
+                    elements = new String[maxElements];
+                }
             }
         }
         List<BufferedReader> readers = new ArrayList<>(slices);
         String[] firstLines = new String[slices];
-        String line;
 
         for (int i = 0; i < slices; i++) {
             readers.add(new BufferedReader(new FileReader(String.format("C:\\csv\\sortedFiles\\%s%d.csv", fileNames, i))));
@@ -46,38 +70,57 @@ public class ExternalSorter {
         int min;
         String[] elements;
         BufferedWriter writer = new BufferedWriter(new FileWriter("C:\\csv\\sortedFiles\\result.csv"));
-
+        int written = 0;
         //Reads all Files
-        for (int i = 0; i < maxElements; i++) {
+        while (true) {
             min = Integer.MAX_VALUE;
             //Reads finds min value for each reader
             for (int k = 0; k < slices; k++) {
-                elements = firstLines[k].split(",");
-                if (min >= Integer.parseInt(elements[1])) {
-                    min = Integer.parseInt(elements[1]);
+                if (firstLines[k] != null) {
+                    elements = firstLines[k].split(",");
+                    if (min >= Integer.parseInt(elements[1])) {
+                        min = Integer.parseInt(elements[1]);
+                    }
                 }
             }
 
             //Writes every value which is == to min and continues onward.
             for (int j = 0; j < slices; j++) {
-                elements = firstLines[j].split(",");
-                if (min <= Integer.parseInt(elements[1])) {
-                    writer.write(firstLines[j]);
-                    writer.newLine();
-                    firstLines[j] = readers.get(j).readLine();
+                if (firstLines[j] != null) {
+                    elements = firstLines[j].split(",");
+                    if (min == Integer.parseInt(elements[1])) {
+                        written++;
+                        writer.write(firstLines[j]);
+                        writer.newLine();
+                        firstLines[j] = readers.get(j).readLine();
+                    }
                 }
             }
+            int count = 0;
+            for (int i = 0; i < slices; i++) {
+                if (firstLines[i] != null) {
+                    break;
+                } else {
+                    count++;
+                }
+
+            }
+            if (count == slices) {
+                for (int i = 0; i < slices; i++) {
+                    readers.get(i).close();
+                    Files.delete(Paths.get(String.format("C:\\csv\\sortedFiles\\%s%d.csv", fileNames, i)));
+                }
+                writer.close();
+                break;
+            }
         }
-        for(int i = 0; i<slices; i++){
-            readers.get(i).close();
-            Files.delete(Paths.get(String.format("C:\\csv\\sortedFiles\\%s%d.csv",fileNames,i)));
-        }
-        writer.close();
-        System.out.printf("TIME: %d", (System.currentTimeMillis()-begin));
+        System.out.println(written);
+        System.out.printf("TIME: %d", (System.currentTimeMillis() - begin));
     }
 
     public static void merge(String[] strings, String[] temp, int from, int mid, int to) {
         int k = from, i = from, j = mid + 1;
+        try {
             while (i <= mid && j <= to) {
                 String[] leftElement = strings[i].split(",");
                 String[] rightElement = strings[j].split(",");
@@ -87,6 +130,9 @@ public class ExternalSorter {
                     temp[k++] = strings[j++];
                 }
             }
+        }catch (NullPointerException e){
+            System.out.println("NULL");
+        }
 
         while (i < strings.length && i <= mid) {
             temp[k++] = strings[i++];
@@ -124,5 +170,3 @@ public class ExternalSorter {
         }
     }
 }
-
-;
